first pass at compiling certain expressions to bytecode and interpreting the bytecode. It compiles, but haven't tested yet

+++ src/symbol.h
@@ -34,7 +34,7 @@ struct SymbolInfo
 		struct UVarData
 		{
 			AstVarDeclStmt * pVarDeclStmt;
-			u32 byteOffset;
+			uintptr byteOffset;
 		} varData;
 
 		struct UFuncData
